Quarantine planar frustum mesh primitive helpers
